+++ apps/docs/src/app/core/component-docs/action-sheet/examples/action-sheet-mobile/action-sheet-mobile-example.component.ts
@@ -11,7 +11,13 @@ export class ActionSheetMobileExampleComponent {
     @ViewChild(ActionSheetComponent)
     actionSheetComponent: ActionSheetComponent;
 
+    isOpen = false;
+
     constructor(private _messageToastService: MessageToastService) {}
+
+    isOpenChange(isOpen: boolean): void {
+        this.isOpen = isOpen;
+    }
 
     actionPicked(action: string): void {
         this.openMessageToast(action);