+++ libs/core/src/lib/notification/notification/notification.component.spec.ts
@@ -25,8 +25,7 @@ class TemplateTestComponent {
 
 @NgModule({
     declarations: [TemplateTestComponent],
-    imports: [CommonModule, BrowserModule, NotificationModule, ButtonModule],
-    entryComponents: [TemplateTestComponent]
+    imports: [CommonModule, BrowserModule, NotificationModule, ButtonModule]
 })
 class TestModule {}