+++ libs/fn/src/lib/form/form-item/form-item.component.ts
@@ -47,7 +47,7 @@ export class FormItemComponent implements CssClassBuilder, OnInit, OnChanges {
 
     _icon: string;
 
-    state: FormStates;
+    state?: FormStates;
 
     constructor(private _elementRef: ElementRef) {}