+++ libs/platform/src/lib/dynamic-page/dynamic-page.component.spec.ts
@@ -129,7 +129,7 @@ describe('DynamicPageComponent default values', () => {
         const throttleTime = 100;
         const contentEl = fixture.debugElement.query(By.css('fd-dynamic-page-content'));
 
-        spyOn(component.dynamicPageHeaderComponent.collapseChange, 'emit');
+        spyOn(component.dynamicPageHeaderComponent.collapsedChange, 'emit');
 
         contentEl.nativeElement.dispatchEvent(new Event('scroll'));
 
@@ -137,7 +137,7 @@ describe('DynamicPageComponent default values', () => {
         tick(throttleTime);
 
         fixture.whenStable().then(() => {
-            expect(component.dynamicPageHeaderComponent.collapseChange.emit).toHaveBeenCalled();
+            expect(component.dynamicPageHeaderComponent.collapsedChange.emit).toHaveBeenCalled();
         });
     }));
 });