+++ libs/platform/src/lib/form/checkbox-group/checkbox-group.component.ts
@@ -8,6 +8,7 @@ import {
     Host,
     Inject,
     Input,
+    isDevMode,
     Optional,
     Output,
     QueryList,
@@ -17,6 +18,8 @@ import {
     ViewEncapsulation
 } from '@angular/core';
 import { NgControl, NgForm } from '@angular/forms';
+import { RangeSelector } from '@fundamental-ngx/core/utils';
+import { SelectionModel } from '@angular/cdk/collections';
 
 import {
     FormField,
@@ -24,17 +27,16 @@ import {
     InLineLayoutCollectionBaseInput,
     RESPONSIVE_BREAKPOINTS_CONFIG,
     ResponsiveBreakPointConfig,
-    ResponsiveBreakpointsService
+    ResponsiveBreakpointsService,
+    coerceArraySafe
 } from '@fundamental-ngx/platform/shared';
-import { CheckboxComponent, PlatformCheckboxChange } from '../checkbox/checkbox.component';
+import { CheckboxComponent } from '../checkbox/checkbox.component';
 
 /**
  * Checkbox group implementation based on the
  * https://github.com/SAP/fundamental-ngx/wiki/Platform:-CheckboxGroup-Technical-Design
  * documents.
- *
  */
-
 @Component({
     selector: 'fdp-checkbox-group',
     templateUrl: './checkbox-group.component.html',
@@ -46,11 +48,13 @@ export class CheckboxGroupComponent extends InLineLayoutCollectionBaseInput {
     /**
      * value for selected checkboxes.
      */
-    get value(): any {
-        return super.getValue();
+    @Input()
+    get value(): any[] {
+        return this.getValue();
     }
-    set value(selectedValue: any) {
-        super.setValue(selectedValue);
+    set value(selectedValue: any[]) {
+        this.setValue(selectedValue);
+        this._updateSelectionModelByValue();
     }
 
     /**
@@ -66,14 +70,20 @@ export class CheckboxGroupComponent extends InLineLayoutCollectionBaseInput {
     }
 
     /**
+     * @deprecated
      * Establishes two way binding, when checkbox group used outside form.
      */
     @Input()
     get checked(): string[] {
-        return this._checked;
+        if (isDevMode()) {
+            console.warn('"checked" is deprecated. Use "value" instead');
+        }
+        return this.value;
     }
     set checked(checkedValues: string[]) {
-        this._checked = checkedValues;
+        if (isDevMode()) {
+            console.warn('"checked" is deprecated. Use "value" instead');
+        }
         this.value = checkedValues;
     }
 
@@ -85,15 +95,22 @@ export class CheckboxGroupComponent extends InLineLayoutCollectionBaseInput {
     @ViewChildren(CheckboxComponent)
     viewCheckboxes: QueryList<CheckboxComponent>;
 
-    @Output()
-    readonly valueChange: EventEmitter<PlatformCheckboxChange> = new EventEmitter<PlatformCheckboxChange>();
-
     /** Emits checked change event */
     @Output()
-    readonly checkedChange: EventEmitter<string[]> = new EventEmitter<string[]>();
+    readonly valueChange: EventEmitter<string[]> = new EventEmitter<string[]>();
 
-    /** @hidden used for two way binding, when used outside form */
-    private _checked: string[];
+    /**
+     * @deprecated - use "valueChange" instead
+     * Emits checked change event
+     */
+    @Output()
+    readonly checkedChange = this.valueChange;
+
+    /** @hidden */
+    private readonly _rangeSelector = new RangeSelector();
+
+    /** @hidden */
+    readonly _selectionModel = new SelectionModel(true);
 
     constructor(
         cd: ChangeDetectorRef,
@@ -118,16 +135,61 @@ export class CheckboxGroupComponent extends InLineLayoutCollectionBaseInput {
     }
 
     /**
-     * raises event when Checkbox group value changes.
-     * @param event: contains checkbox and event in a PlatformCheckboxChange class object.
+     * @hidden
+     * @param selectionState actual value of the checkbox
+     * @param checkboxTrueValue thuthy value to compare the "selectionState" with
      */
-    public groupValueChanges(event: PlatformCheckboxChange): void {
-        if (event instanceof PlatformCheckboxChange) {
-            this.checked = event.checked;
-            this.onTouched();
-            this.valueChange.emit(event);
-            this.checkedChange.emit(this.checked);
+    onModelChanged(selectionState: any, checkboxTrueValue: any): void {
+        const isSelected = selectionState === checkboxTrueValue;
+        if (isSelected) {
+            this._selectionModel.select(checkboxTrueValue);
+        } else {
+            this._selectionModel.deselect(checkboxTrueValue);
         }
+        this._setValueBySelectionModel();
+        this.onTouched();
+        this.valueChange.emit(this.value);
+    }
+
+    /** @hidden */
+    onCheckboxClick(index: number, source: 'contentChildren' | 'list', event?: PointerEvent): void {
+        // this handler will be invoked after "valueChange"
+        const queryList = source === 'list' ? this.viewCheckboxes : this.contentCheckboxes;
+        const target = queryList.get(index);
+
+        const isChecked = this._selectionModel.isSelected(target.values.trueValue);
+        this._rangeSelector.onRangeElementToggled(index, event);
+
+        if (!this._rangeSelector.lastRangeSelectionState.isRangeSelection) {
+            // no need to proceed, if it's not a range selection - current checkbox is already updated
+            return;
+        }
+
+        this._rangeSelector.applyValueToEachInRange((idx) => {
+            const checkbox = queryList.get(idx);
+            if (checkbox.disabled) {
+                return;
+            }
+            const checkboxValue = checkbox.values.trueValue;
+            if (isChecked) {
+                this._selectionModel.select(checkboxValue);
+            } else {
+                this._selectionModel.deselect(checkboxValue);
+            }
+        });
+        this._setValueBySelectionModel();
+        this.valueChange.emit(this.value);
+    }
+
+    /** @hidden */
+    writeValue(selectedValue: any): void {
+        super.writeValue(coerceArraySafe(selectedValue));
+        this._updateSelectionModelByValue();
+    }
+
+    /** @inheritdoc */
+    protected setValue(value: any, emitOnChange = true): void {
+        super.setValue(coerceArraySafe(value), emitOnChange);
     }
 
     /**
@@ -147,5 +209,18 @@ export class CheckboxGroupComponent extends InLineLayoutCollectionBaseInput {
     /** @hidden */
     public getListItemDisabledValue(item: CheckboxGroupComponent['list'][number]): boolean {
         return this.disabled || typeof item === 'string' ? this.disabled : item.disabled;
+    }
+
+    /** @hidden */
+    private _setValueBySelectionModel(): void {
+        super.setValue(this._selectionModel.selected);
+        this._cd.markForCheck();
+    }
+
+    /** @hidden */
+    private _updateSelectionModelByValue(): void {
+        this._selectionModel.clear();
+        this.value.forEach((v) => this._selectionModel.select(v));
+        this._cd.markForCheck();
     }
 }