+++ libs/platform/src/lib/shared/pipes/convert-bytes/convert-bytes.pipe.spec.ts
@@ -47,16 +47,19 @@ describe('Pipe: ConvertBytesPipe', () => {
     });
 
     it('should convert bytes to exabyte label', () => {
+        // eslint-disable-next-line @typescript-eslint/no-loss-of-precision
         const result = pipe.transform(4684567346346346345);
         expect(result).toEqual('4.06 EB');
     });
 
     it('should convert bytes to zettabyte label', () => {
+        // eslint-disable-next-line @typescript-eslint/no-loss-of-precision
         const result = pipe.transform(5685685673452523537484);
         expect(result).toEqual('4.82 ZB');
     });
 
     it('should convert bytes to yottabyte label', () => {
+        // eslint-disable-next-line @typescript-eslint/no-loss-of-precision
         const result = pipe.transform(5685685673452523537484568);
         expect(result).toEqual('4.7 YB');
     });